requirements.txt
selenium
webdriver-manager
httpx
python-dotenv